SBERDOMA-762 New user register screen markup
- major layout refactoring
- DRY
- KISS
