+++ apps/condo/domains/common/components/containers/BaseLayout/components/styles.ts
@@ -9,34 +9,37 @@ export const SIDE_MENU_WIDTH = 256
 
 // Ant inputs
 export const formInputFixCss = css`
-.ant-input-affix-wrapper:focus,
-.ant-input-affix-wrapper-focused,
-.ant-form-item-has-error .ant-input-affix-wrapper:focus,
-.ant-form-item-has-error .ant-input-affix-wrapper-focused {
+  body {
+    height: 100%
+  }
+  .ant-input-affix-wrapper:focus,
+  .ant-input-affix-wrapper-focused,
+  .ant-form-item-has-error .ant-input-affix-wrapper:focus,
+  .ant-form-item-has-error .ant-input-affix-wrapper-focused {
     background-color: ${colors.white};
-}
-.ant-input-affix-wrapper > input.ant-input,
-.ant-form-item-has-error .ant-input-affix-wrapper > input.ant-input {
+  }
+  .ant-input-affix-wrapper > input.ant-input,
+  .ant-form-item-has-error .ant-input-affix-wrapper > input.ant-input {
     -webkit-box-shadow: inset 0 0 0 50px ${colors.ultraLightGrey};
     -webkit-text-fill-color: ${colors.black};
-}
-.ant-form-item-has-error .ant-input-affix-wrapper > input.ant-input:focus,
-.ant-input-affix-wrapper > input.ant-input:focus {
+  }
+  .ant-form-item-has-error .ant-input-affix-wrapper > input.ant-input:focus,
+  .ant-input-affix-wrapper > input.ant-input:focus {
     -webkit-box-shadow: inset 0 0 0 50px ${colors.white};
     -webkit-text-fill-color: ${colors.black};
     background-color: ${colors.white};
-}
-.ant-form-item-has-error .ant-input-affix-wrapper:not(.ant-input-affix-wrapper-disabled):hover {
+  }
+  .ant-form-item-has-error .ant-input-affix-wrapper:not(.ant-input-affix-wrapper-disabled):hover {
     border-color: ${red[5]};
-}
-.ant-form-item-has-error .ant-input-affix-wrapper-focused:hover,
-.ant-form-item-has-error .ant-input-affix-wrapper:focus:hover {
+  }
+  .ant-form-item-has-error .ant-input-affix-wrapper-focused:hover,
+  .ant-form-item-has-error .ant-input-affix-wrapper:focus:hover {
     border-color: ${red[5]};
     background-color: ${colors.white};
-}
-.ant-input:focus:hover, .ant-input:active:hover {
+  }
+  .ant-input:focus:hover, .ant-input:active:hover {
     background-color: ${colors.white};
-}
+  }
 `
 
 const shadow = css`
