+++ apps/condo/domains/user/components/auth/InputPhoneForm.tsx
@@ -1,0 +1,164 @@
+import { Button } from '@condo/domains/common/components/Button'
+import { PhoneInput } from '@condo/domains/common/components/PhoneInput'
+import { runMutation } from '@condo/domains/common/utils/mutations.utils'
+import { normalizePhone } from '@condo/domains/common/utils/phone'
+import { getClientSideSenderInfo } from '@condo/domains/common/utils/userid.utils'
+import { TOO_MANY_REQUESTS } from '@condo/domains/user/constants/errors'
+import { START_CONFIRM_PHONE_MUTATION } from '@condo/domains/user/gql'
+import { useMutation } from '@core/next/apollo'
+import { useIntl } from '@core/next/intl'
+import { Col, Form, Row, Tooltip, Typography } from 'antd'
+import Router from 'next/router'
+import React, { useCallback, useContext, useMemo, useState } from 'react'
+import { FormattedMessage } from 'react-intl'
+import { RegisterContext } from './RegisterContextProvider'
+
+const FORM_LAYOUT = {
+    labelCol: { span: 10 },
+    wrapperCol: { span: 14 },
+}
+
+interface IInputPhoneFormProps {
+    onFinish: () => void
+}
+
+export const InputPhoneForm: React.FC<IInputPhoneFormProps> = ({ onFinish })=> {
+    const [form] = Form.useForm()
+    const intl = useIntl()
+    const PhoneMsg = intl.formatMessage({ id: 'pages.auth.register.field.Phone' })
+    const RegisterHelpMessage = intl.formatMessage({ id: 'pages.auth.reset.RegisterHelp' })
+    const UserAgreementFileName = intl.formatMessage({ id: 'pages.auth.register.info.UserAgreementFileName' })
+    const ExamplePhoneMsg = intl.formatMessage({ id: 'example.Phone' })
+    const FieldIsRequiredMsg = intl.formatMessage({ id: 'FieldIsRequired' })
+    const SMSTooManyRequestsError = intl.formatMessage({ id: 'pages.auth.TooManyRequests' })
+    const RegisterMsg = intl.formatMessage({ id: 'Register' })
+    const SberIdRegisterMsg = intl.formatMessage({ id: 'SberIdRegister' })
+    const NotImplementedYetMessage = intl.formatMessage({ id: 'NotImplementedYet' })
+
+    const { setToken, setPhone, handleReCaptchaVerify } = useContext(RegisterContext)
+    const [smsSendError, setSmsSendError] = useState(null)
+    const [isLoading, setIsLoading] = useState(false)
+    const [startPhoneVerify] = useMutation(START_CONFIRM_PHONE_MUTATION)
+    const ErrorToFormFieldMsgMapping = useMemo(() => {
+        return {
+            [TOO_MANY_REQUESTS]: {
+                name: 'phone',
+                errors: [SMSTooManyRequestsError],
+            },
+        }
+    }, [intl])
+
+    const startConfirmPhone = useCallback(async () => {
+        const registerExtraData = {
+            dv: 1,
+            sender: getClientSideSenderInfo(),
+        }
+        const { phone: inputPhone } = form.getFieldsValue(['phone'])
+        const phone = normalizePhone(inputPhone)
+        setPhone(phone)
+        const captcha = await handleReCaptchaVerify('start_confirm_phone')
+        const variables = { data: { ...registerExtraData, phone, captcha } }
+        setIsLoading(true)
+
+        // @ts-ignore TODO(Dimitreee): remove after runMutation typo
+        return runMutation({
+            mutation: startPhoneVerify,
+            variables,
+            onCompleted: (data) => {
+                const { data: { result: { token } } } = data
+                setToken(token)
+                Router.push(`/auth/register?token=${token}`)
+                onFinish()
+            },
+            onFinally: () => {
+                setIsLoading(false)
+            },
+            intl,
+            form,
+            ErrorToFormFieldMsgMapping,
+        }).catch(() => {
+            setIsLoading(false)
+        })
+    }, [intl, form])
+
+    return (
+        <Form
+            {...FORM_LAYOUT}
+            form={form}
+            name='register-input-phone'
+            onFinish={startConfirmPhone}
+            colon={false}
+            labelAlign='left'
+            requiredMark={false}
+        >
+            <Row gutter={[0, 60]}>
+                <Col span={24}>
+                    <Row gutter={[0, 40]}>
+                        <Col span={24}>
+                            <Typography.Paragraph>{RegisterHelpMessage}</Typography.Paragraph>
+                        </Col>
+                        <Col span={24}>
+                            <Form.Item
+                                name='phone'
+                                label={PhoneMsg}
+                                rules={[
+                                    {
+                                        required: true,
+                                        message: FieldIsRequiredMsg,
+                                    },
+                                    () => ({
+                                        validator () {
+                                            if (!smsSendError) {
+                                                return Promise.resolve()
+                                            }
+                                            return Promise.reject(smsSendError)
+                                        },
+                                    }),
+                                ]}
+                            >
+                                <PhoneInput placeholder={ExamplePhoneMsg} onChange={() => setSmsSendError(null)}/>
+                            </Form.Item>
+                        </Col>
+                        <Col span={24}>
+                            <FormattedMessage
+                                id='pages.auth.register.info.UserAgreement'
+                                values={{
+                                    link: (
+                                        <Button type={'inlineLink'} size={'small'} target='_blank' href={'/policy.pdf'} rel='noreferrer'>
+                                            {UserAgreementFileName}
+                                        </Button>
+                                    ),
+                                }}
+                            />
+                        </Col>
+                    </Row>
+                </Col>
+                <Col span={24}>
+                    <Row justify={'space-between'} gutter={[0, 12]}>
+                        <Col lg={24} xl={11}>
+                            <Button
+                                key='submit'
+                                type='sberPrimary'
+                                htmlType='submit'
+                                loading={isLoading}
+                            >
+                                {RegisterMsg}
+                            </Button>
+                        </Col>
+                        <Col lg={24} xl={13}>
+                            <Tooltip title={NotImplementedYetMessage}>
+                                <Button
+                                    secondary
+                                    key='submit'
+                                    type='sberPrimary'
+                                >
+                                    {SberIdRegisterMsg}
+                                </Button>
+                            </Tooltip>
+                        </Col>
+                    </Row>
+                </Col>
+            </Row>
+        </Form>
+    )
+}
