+++ apps/condo/domains/user/components/auth/hooks.ts
@@ -1,0 +1,66 @@
+import { Rule } from 'rc-field-form/lib/interface'
+import { useMemo } from 'react'
+import { MIN_PASSWORD_LENGTH } from '@condo/domains/user/constants/common'
+import { useIntl } from '@core/next/intl'
+
+type ValidatorsMap = {
+    [key: string]: Rule[]
+}
+
+export const useRegisterFormValidators = () => {
+    const intl = useIntl()
+    const PleaseInputYourEmailMsg = intl.formatMessage({ id: 'pages.auth.PleaseInputYourEmail' })
+    const PleaseInputYourPasswordMsg = intl.formatMessage({ id: 'pages.auth.PleaseInputYourPassword' })
+    const PleaseInputYourNameMsg = intl.formatMessage({ id: 'pages.auth.PleaseInputYourName' })
+    const EmailIsNotValidMsg = intl.formatMessage({ id: 'pages.auth.EmailIsNotValid' })
+    const PleaseConfirmYourPasswordMsg = intl.formatMessage({ id: 'pages.auth.PleaseConfirmYourPassword' })
+    const TwoPasswordDontMatchMsg = intl.formatMessage({ id: 'pages.auth.TwoPasswordDontMatch' })
+    const PasswordIsTooShortMsg = intl.formatMessage({ id: 'pages.auth.PasswordIsTooShort' })
+
+    return useMemo<ValidatorsMap>(() => {
+        return {
+            phone: [{ required: true }],
+            name: [
+                {
+                    required: true,
+                    message: PleaseInputYourNameMsg,
+                    whitespace: true,
+                },
+            ],
+            email: [
+                {
+                    type: 'email',
+                    message: EmailIsNotValidMsg,
+                },
+                {
+                    required: true,
+                    message: PleaseInputYourEmailMsg,
+                },
+            ],
+            password: [
+                {
+                    required: true,
+                    message: PleaseInputYourPasswordMsg,
+                },
+                {
+                    min: MIN_PASSWORD_LENGTH,
+                    message: PasswordIsTooShortMsg,
+                },
+            ],
+            confirm: [
+                {
+                    required: true,
+                    message: PleaseConfirmYourPasswordMsg,
+                },
+                ({ getFieldValue }) => ({
+                    validator (_, value) {
+                        if (!value || getFieldValue('password') === value) {
+                            return Promise.resolve()
+                        }
+                        return Promise.reject(TwoPasswordDontMatchMsg)
+                    },
+                }),
+            ],
+        }
+    }, [intl])
+}
