+++ apps/condo/lang/en.json
@@ -358,6 +358,7 @@
   "ReadMore": "Expand",
   "Redirecting": "Redirecting ...",
   "Register": "Register",
+  "SberIdRegister": "Регистрация через Сбер ID",
   "Reject": "Reject",
   "ResendVerificationCode": "Resend verification code",
   "Save": "Save",
