+++ apps/condo/lang/ru.json
@@ -127,7 +127,7 @@
   "pages.auth.EmailIsNoFound": "Email не найден",
   "pages.auth.EmailIsNotRegistered": "Email еще не зарегистрирован",
   "pages.auth.EmailIsNotValid": "Email имеет странный формат",
-  "pages.auth.FooterText": "Проблемы с регистрацией или входом? Напишите на {email} или позвоните по телефону {phone}",
+  "pages.auth.FooterText": "Проблемы с регистрацией или входом?\nНапишите на {email} или позвоните по телефону {phone}",
   "pages.auth.ChangePasswordTokenErrorLabel": "Ссылка на восстановление пароля больше не работает",
   "pages.auth.ChangePasswordTokenErrorMessage": "Вы можете инициировать восстановление пароля ещё раз",
   "pages.auth.ChangePasswordTokenErrorConfirmLabel": "Восстановить пароль",
@@ -358,6 +358,7 @@
   "ReadMore": "Раскрыть",
   "Redirecting": "Перенаправляем ...",
   "Register": "Зарегистрироваться",
+  "SberIdRegister": "Регистрация через Сбер ID",
   "Reject": "Отклонить",
   "ResendVerificationCode": "Повторно отправить код",
   "Save": "Сохранить",
